Remove the system cscope and rebuild it ourselves.
This will bring us from version 15.8 -> 15.9.
But more importantly, we have stole the patches from the Ubuntu cscope
deb, patches which fix the problem of cscope not recognizing functions
which take functions as arguments.
This fix adds a lot of functions in the Linux kernel to the index that
cscope produces.
Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.