Add `add_user_to_sudoers` crust app
This is a program which is intended to be ran in the Docker
container upon before anything else is done. It must be installed
such that:
    * It is owned by root
    * Its SUID bit is set

Its purpose is when one has mapped in `/etc/passwd` and `/etc/group`
to the Docker container and has set the user to themselves via
`docker run -u $(id -u) ...`. This app will take the group
of the user and add it to the sudoers file (via a new file in
/etc/sudoers.d/) so that one may use `sudo` correctly.