Fix permissions bit on Docker COPY(ed) file
It turns out that Docker's `COPY` directive in
the Dockerfile has a `--chown` option, which is good,
but they still do not have a `--chmod` flag, which
has been a request since 2017!
Because of this, the previous Docker image version
didn't get the `add_user_to_sudoers` application installed
correctly with its SUID bit set. This commit fixes that
by unfortunately having to add an additional `RUN` directive.